Move Debug Screen in the components folder
