+++ src/RootNavigation/RootBottomNavigator/RootBottomNavigator.screens.ts
@@ -6,7 +6,7 @@
 import { ComponentType } from 'react';
 
 import RNDemoApp from '../../RNDemoApp/RNDemoApp';
-import { DebugScreenContainer } from '../../DebugScreen/DebugScreenContainer';
+import { DebugScreenContainer } from '../../components/DebugScreen/DebugScreenContainer';
 import { RootBottomNavigatorRoutes } from './RootBottomNavigator.routes';
 import { HomeStackNavigator } from '../../HomeStackNavigator/HomeStackNavigator';
