+++ src/components/DebugScreen/DebugScreen.tsx
