+++ src/components/DebugScreen/LabelValueText.tsx
